requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0\
